Add files via upload
Uses 2D Nav Goal in rviz to go to picked location. 
Note:  Running this does move robot to location and also the specified angle. This does not shut down properly so it continues to run. Robot does stop when it reaches goal but ctr + c has to be used to exit for now.